cmd: Let the terminal know the active container on all host OSes
Prompt is a new terminal emulation application [1] designed for a
container-oriented desktop that implements the OSC 777 escape sequence
to track and preserve the user's current Toolbx container [2].  Hence,
Fedora's fork of GNOME Terminal is no longer the only one to offer this.

The implementation in Prompt is already better because it has a
user-visible setting to disable this integration with Toolbx, in case
the user doesn't want it.  Therefore, it's time to let users of all host
operating systems enjoy this feature.

This reverts commits a3e8d8d12bac6dd63010b71c6e091486fb585f37 and
3a96feba478448cd0c5628238619c3805c6aca1e.

[1] https://gitlab.gnome.org/chergert/prompt

[2] https://gitlab.freedesktop.org/terminal-wg/specifications/-/issues/17

https://github.com/containers/toolbox/issues/218