test/system: Test that the HOSTNAME environment variable is set
Bash automatically sets the HOSTNAME environment variable to the name of
the current host [1] as returned by gethostname(2), which is the same as
hostname(1).

However, on Fedora, from Fedora 33 onwards, /etc/profile sets the
HOSTNAME environment variable to 'hostnamectl --transient' [2], and,
from Fedora 35 onwards, it has a fallback to hostname(1) [3].  These two
approaches return different values when used inside a Toolbx container.
The former picks up the hostname of the host operating system, while the
fallback gets the name that was set when creating the container with
'podman create --hostname toolbox ...'.

Hence, the value of HOSTNAME inside a Toolbx container for Fedora
depends on whether the corresponding version of the fedora-toolbox image
contained hostnamectl(1) or not.

[1] https://www.gnu.org/software/bash/manual/html_node/Bash-Variables.html

[2] setup commit eb9cc4dce89be24f
    https://pagure.io/setup/c/eb9cc4dce89be24f
    https://bugzilla.redhat.com/show_bug.cgi?id=1745245

[3] setup commit ddd74b5d971a734c
    https://pagure.io/setup/c/ddd74b5d971a734c
    https://pagure.io/setup/pull-request/28
    https://bugzilla.redhat.com/show_bug.cgi?id=1938223

https://github.com/containers/toolbox/issues/558