build: Do not use use auto dependencies for shell completion scripts
It is difficult for downstream to track dependencies when they are
automatic. This results in shell completion scripts not being
installed in packages because builders do not have the right
dependency. This commit adds meson feature arguments to guard those
dependencies so that downstream distributions can use
`-Dauto_features=enabled`.

For more explanation, see rule #3 of:
https://blogs.gnome.org/mcatanzaro/2022/07/15/best-practices-for-build-options/

https://github.com/containers/toolbox/pull/1442

Signed-off-by: Valentin David <me@valentindavid.com>