images/fedora: Tweak the 'summary' label for Fedora 38 and 39
The concept of a 'base' image for creating Toolbx containers hasn't
existed since Toolbx 0.0.10, when the Buildah dependency and the
user-specific customized image were dropped [1].  That was a long time
ago.  Toolbx containers created before that were never supported by the
Go implementation [2].

Therefore, it's time to update the terminology.

[1] Commit 8b84b5e4604921fa
    https://github.com/containers/toolbox/commit/8b84b5e4604921fa
    https://github.com/containers/toolbox/pull/160

[2] Commit 238f2451e7d7d54a
    https://github.com/containers/toolbox/commit/238f2451e7d7d54a
    https://github.com/containers/toolbox/pull/318

https://github.com/containers/toolbox/pull/1456