images/fedora: Add all the iconv converter modules for glibc
Currently, the glibc package mentions glibc-gconv-extra, which contains
the iconv converter modules that are not commonly used, as a weak
dependency [1].  However, the glibc package is part of the fedora base
image, which doesn't include weak dependencies [2], and that leads to
glibc-gconv-extra going missing.

Only the images for currently maintained Fedoras (ie., 38 and 39) were
updated.

[1] Fedora glibc commit 51ca5d3334c3cd35
    https://src.fedoraproject.org/rpms/glibc/c/51ca5d3334c3cd35
    https://src.fedoraproject.org/rpms/glibc/pull-request/36
    https://bugzilla.redhat.com/show_bug.cgi?id=1812191

[2] fedora-kickstarts commit 1c39c0adb0d44866
    https://pagure.io/fedora-kickstarts/c/1c39c0adb0d44866
    https://pagure.io/fedora-kickstarts/pull-request/551
    https://pagure.io/releng/issue/8530

https://github.com/containers/toolbox/pull/1456