.zuul: Enable testing on Fedora 40
The current timeout of 1 hour that's used for stable Fedoras has proved
insufficient for Fedora 40.  It's not clear why that is.  It's possible
that since Fedora 40 is currently Branched and pre-release Fedoras use
Linux kernels that are built with debugging enabled, it's slower than
stable Fedoras.

Therefore, the same timeout as Fedora Rawhide has been used.

https://github.com/containers/toolbox/pull/1468