cmd/list, pkg/podman: Limit access to the raw 'podman ps' JSON
This builds on top of commit 56fcb0b4d26adb9c.

The toolboxContainer type has been renamed to Container and moved into
the podman package.

There is nothing Toolbx specific about the type - it represents any
container returned by 'podman ps'.  The containers are only later
filtered for Toolbx containers.

Secondly, having the Container type inside the podman package makes it
possible to encapsulate the unmarshalling of the JSON within the package
without exposing the raw JSON to outside consumers.  This is desirable
because the unmarshalling involves tracking changes in the JSON output
by different Podman versions, and it's better to limit such details to
the podman package.

https://github.com/containers/toolbox/pull/1485