cmd/create, pkg/podman: Let 'create' use an image without a name
Currently, it's not possible to create a Toolbx container from an image
without a name:
  $ podman build --squash images/fedora/f39
  STEP 1/21: FROM registry.fedoraproject.org/fedora:39
  STEP 2/21: ARG NAME=fedora-toolbox
  STEP 3/21: ARG VERSION=39
  ...
  --> 2f9bdf11c8d4
  2f9bdf11c8d4d7674dfb17d8edcfd13475d8636077f1a6208ecd616de77d7f80
  $ toolbox create --image 2f9bdf11c8d4
  Error: empty RepoTag for image 2f9bdf11c8d4

The image's fully qualified name is fetched from its RepoTags for purely
cosmetic reasons to show a precise human-readable name in the debug logs
and 'podman inspect --type container'.  Therefore, there's no reason to
fail the creation of a Toolbx container in the absence of it.

Note that an image without a name will have an empty RepoTags array in
the JSON returned by 'podman inspect --format json --type image'.  It's
different from not having a RepoTags array at all in the JSON, which may
or may not be indicative of a more serious problem and will continue to
fail the creation of the Toolbx container as before.

https://github.com/containers/toolbox/pull/1486