cmd/run, pkg/podman: Split Inspect() into two for containers and images
In future, it will be good if podman.Inspect() returned a Container or
Image object instead of a []map[string]interface{} that the caller has
to parse.  This is because parsing the []map[string]interface{} involves
tracking changes in the JSON output by different Podman versions, and
it's better to limit such details to the podman package.

Splitting podman.Inspect() into two separate functions for containers
and images is one way of achieving that.

https://github.com/containers/toolbox/pull/1487