cmd/run: Show the entry point's debug logs & errors in 'enter' & 'run'
Until now, if the entry point of a Toolbx container encountered an
error, while starting the container as part of the 'enter' and 'run'
commands, the specific error wouldn't be presented to the user by those
commands.  Instead, the user would have to use 'podman start --attach'
or 'podman logs' to retrieve it.  Same for the debug logs coming from
the entry point.

The lack of relevant information and insight made it difficult for users
to debug their containers or file high quality bug reports.

This is addressed by using 'podman logs', as part of the 'enter' and
'run' commands to fetch everything from the entry point's standard error
and output streams, which means both debug logs and errors, separating
them out, and presenting them to the user depending on the chosen debug
or verbosity level.

The debug logs from 'podman logs' are in the logfmt format [1], because
that's the default behaviour of Logrus when a terminal device is not
attached [2].  Logs without a 'level' key are assumed to be errors.

It's assumed that the 'podman logs' process can crash or terminate,
because the entry point crashed or got killed due to an out-of-band
'podman stop' or encountered an error.  Under such circumstances, the
'enter' and 'run' commands will terminate immediately if they have
already read any error coming from the entry point, if not then they
will wait for the timeout.

If the entry point successfully initializes the Toolbx container, then
'enter' and 'run' will cancel the 'podman logs' process, parse and show
any pending logs, and then terminate.

It's possible to detect the creation of the initialization stamp file
before all that was written by the entry point has been read from
'podman logs', causing all or some of the debug logs to be not shown as
part of 'enter' and 'run'.  This is because the creation of the
initialization stamp file is detected by a quick inotify(7) watch within
the 'enter' and 'run' processes, while the logs slowly change hands
across multiple entities - from the entry point to conmon(8) to the
systemd journal to 'podman logs'.

[1] https://brandur.org/logfmt

[2] https://pkg.go.dev/github.com/sirupsen/logrus#section-readme

https://github.com/containers/toolbox/issues/750