cmd, test/system: Support intentially failing the entry point for tests
A subsequent commit will show the debug logs and errors from a Toolbx
container's entry point as part of the 'enter' and 'run' commands.  To
test this behaviour, it will be necessary to intentially fail the entry
point.

Moreover, container start-up is a concurrent operation.  If the entry
point fails too early, then it will be caught by the 'podman inspect'
right after the 'podman start' before the inotify(7) watches are put in
place.  Otherwise, it will be handled by the timeout.  Therefore, it
will be necessary to shake out any bugs arising out of unexpected races.

To address this, two environment variables have been introduced:
  * TOOLBX_DELAY_ENTRY_POINT
  * TOOLBX_FAIL_ENTRY_POINT

The TOOLBX_DELAY_ENTRY_POINT environment variable can be set to a
positive integer during the 'create' command to add a delay, in terms of
seconds, when the Toolbx container's entry point is started by 'enter'
and 'run'.

Similarly, if the TOOLBX_FAIL_ENTRY_POINT environment variable is set to
a positive integer during the 'create' command, the entry point will
later fail during 'enter' and 'run'.  The error message will have only
one line if its value is one, else it will have two.

https://github.com/containers/toolbox/issues/750