test/system: Rewrite a test using the standard Bats helpers
The test was earlier rewritten in commit b0beb68255374392 with custom
code in the hope that it would make it more reliable.  The test has
proven to be reliable in recent times, and the cause for its past
unreliability is unclear.  Therefore, it will be better to remove the
custom code in favour of the standard Bats helpers for the sake of
consistency and readability.

Until Bats 1.10.0, 'run --keep-empty-lines' had a bug where it counted
the trailing newline on the last line as a separate line [1].  However,
Bats 1.10.0 is only available in Fedora >= 39 and is absent from Fedora
38.

[1] Bats commit 6648e2143bffb933
    https://github.com/bats-core/bats-core/commit/6648e2143bffb933
    https://github.com/bats-core/bats-core/issues/708

https://github.com/containers/toolbox/pull/1506