initContainer: also create mount points for sockets
When a socket is bind-mounted to the container, also create a file
mount point for it. Nvidia CDI on the proprietary driver added a
socket for `nvidia-persistence.service` which was failing to be mounted
in the container as no mount point existed.

More logs in the issue below.
Fixes https://github.com/containers/toolbox/issues/1572