build: Bump github.com/briandowns/spinner to 1.18.1
Note that github.com/briandowns/spinner 1.18.1 introduced an undesired
dependency on github.com/mattn/go-isatty for the IsTerminal() API, which
was later removed in 1.23.1 [1].  Fewer dependencies are always good
because it reduces the amount of code in use.

Therefore, this is a step towards using github.com/briandowns/spinner
1.23.1.  Instead of bumping it straight to its final desired version,
doing it in smaller steps makes it easier to bisect any uncaught
regressions in future.

The src/go.sum file was updated with 'go mod tidy'.

[1] github.com/briandowns/spinner commit 8f269dd04fbfe236
    https://github.com/briandowns/spinner/commit/8f269dd04fbfe236
    https://github.com/briandowns/spinner/pull/156

https://github.com/containers/toolbox/pull/1584