test/system: Support host operating systems without VERSION_ID
The VERSION_ID field in os-release(5) is optional [1].  It's absent on
Arch Linux, which follows a rolling-release model and uses the BUILD_ID
field instead:
  BUILD_ID=rolling

A subsequent commit will run the CI on Arch Linux.  Hence, the code to
get the default release from the host operating system can no longer
assume the presence of the VERSION_ID field in os-release(5).

Note that the arch-toolbox image is tagged with 'latest', in accordance
with OCI conventions, not 'rolling' [2,3], which is the os-release(5)
BUILD_ID.

A similar change was made to toolbox(1) in commits 2ee82affebd49a86 and
d14fd7bb50f0e71b.

[1] https://www.freedesktop.org/software/systemd/man/os-release.html

[2] Commit 2568528cb7f52663
    https://github.com/containers/toolbox/commit/2568528cb7f52663
    https://github.com/containers/toolbox/pull/861

[3] Commit a4e5861ae5c93625
    https://github.com/containers/toolbox/commit/a4e5861ae5c93625
    https://github.com/containers/toolbox/pull/1308

https://github.com/containers/toolbox/issues/1438
https://github.com/containers/toolbox/pull/1535