playbooks: Use pkgconfig(bash-completion) to install bash-completion.pc
Unlike Fedora, where the package providing bash-completion.pc changed in
Fedora 41 onwards [1,2], it hasn't changed in CentOS Stream 9.  So, this
doesn't change the behaviour of the CI.

However, the Fedora Packaging Guidelines recommend using pkgconfig(...)
to depend on a *.pc file [3], instead of the name of the package, and
there has been a problem on Fedora caused by not doing that.  Therefore,
it's better to follow the guidelines for consistency and safety.

[1] Fedora bash-completion commit d1f5dc48c0440cc6
    https://src.fedoraproject.org/rpms/bash-completion/c/d1f5dc48c0440cc6
    https://bugzilla.redhat.com/show_bug.cgi?id=1457164

[2] Commit 204919917e4af528
    https://github.com/containers/toolbox/commit/204919917e4af528

[3] https://docs.fedoraproject.org/en-US/packaging-guidelines/PkgConfigBuildRequires/

https://github.com/containers/toolbox/pull/1593