.github/workflows, test/system: Enable 210-ulimit.bats on Ubuntu 22.04
When 'toolbox run ulimit' is invoked, it comes down to invoking
"bash --login -c 'exec ulimit'" inside the container, and it leads to
different outcomes on Fedora and Ubuntu hosts.

On Fedora:
  $ bash --login -c 'exec ulimit'
  unlimited

On Ubuntu:
  $ bash --login -c 'exec ulimit'
  bash: line 1: exec: ulimit: not found

This is because Bash's exec built-in cannot execute another built-in and
needs an external command; and Fedora ships a wrapper for the ulimit
built-in as an external command [1] to satisfy POSIX [2]:
  $ cat /usr/bin/ulimit
  #!/usr/bin/sh
  builtin ulimit "$@"

... that Ubuntu doesn't.

Wrapping the 'ulimit' as 'bash -c ulimit' solves this problem because
then it becomes "bash --login -c 'exec bash -c ulimit'", and the exec
built-in can execute the external command bash(1).

[1] Fedora bash commit 3b09d0d67fe7ff4c
    Fedora bash commit a28ab9933095eaf6
    https://src.fedoraproject.org/rpms/bash/c/3b09d0d67fe7ff4c
    https://src.fedoraproject.org/rpms/bash/c/a28ab9933095eaf6
    https://bugzilla.redhat.com/show_bug.cgi?id=1297166

[2] http://pubs.opengroup.org/onlinepubs/9699919799/utilities/toc.html

https://github.com/containers/toolbox/pull/1599

Signed-off-by: Penn Bauman <me@pennbauman.com>